Removed code to learn about delegates
