+++ TaskCompleter/TaskCompleter/DetailTaskViewController.swift
@@ -11,9 +11,6 @@ import UIKit
 class DetailTaskViewController: UIViewController {
     @IBOutlet var taskLabel: UILabel!
     @IBOutlet var taskDetails: UITextView!
-    var currentTask = String()
-    var taskStorage = TaskStorage.sharedTaskStorage
-    var taskCompleterModelIndex = 0
     
     
     override func viewDidLoad() {
@@ -24,16 +21,14 @@ class DetailTaskViewController: UIViewController {
     }
     
     private func setupDetailVIewUI() {
-        taskLabel?.text = currentTask
+     
         let saveButton = UIBarButtonItem.init(barButtonSystemItem: .save, target: self, action: #selector(saveTaskDetails))
         navigationItem.rightBarButtonItem = saveButton
-        taskStorage.loadTasks()
-        taskDetails.text = taskStorage.taskDataSource[taskCompleterModelIndex].taskDetail
+    
     }
     
     @objc private func saveTaskDetails() {
-        taskStorage.taskDataSource[taskCompleterModelIndex].taskDetail = taskDetails.text
-        taskStorage.saveTasks(taskDataSource: [taskStorage.taskDataSource[taskCompleterModelIndex]])
+        
     }
     
     @objc func dismissKeyboard() {
